move about page sections to src/components/page-about
