+++ src/components/about-page/SectionTeam.js
