+++ src/components/about-page/Statistics.js
