+++ src/pages/About.js
@@ -1,8 +1,8 @@
 
 import Clients from "../components/Clients"
-import Statistics from "../components/product-page/Statistics"
-import SectionTeam from "../components/product-page/SectionTeam";
-import SectionText from "../components/product-page/SectionText";
+import Statistics from "../components/about-page/Statistics"
+import SectionTeam from "../components/about-page/SectionTeam";
+import SectionText from "../components/about-page/SectionText";
 
 export default function About() {
     return (
